// src/modules/ping/enums/ping-status.enum.ts
export enum PingStatusEnum {
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  REJECTED = 'REJECTED',
  IN_PROGRESS = 'IN_PROGRESS',
  PINGER_CONFIRMED = 'PINGER_CONFIRMED',
  PINGED_CONFIRMED = 'PINGED_CONFIRMED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
  DISPUTED = 'DISPUTED',
}