// src/modules/wallet/enums/income-type.enum.ts
export enum IncomeTypeEnum {
  PING_RECEIVED = 'PING_RECEIVED',
  PING_SENT = 'PING_SENT',
  EVENT_JOIN = 'EVENT_JOIN',
  EXCHANGE = 'EXCHANGE',
  REFERRAL = 'REFERRAL',
  SIGNUP_BONUS = 'SIGNUP_BONUS',
  TAG_PURCHASE = 'TAG_PURCHASE',
  SPONSORSHIP = 'SPONSORSHIP',
  BUNDLE_PURCHASE = 'BUNDLE_PURCHASE',
}